up tabulate info : use terminal size in order to avoid a screen overflow
