+++ toolboxes/feel/feelmodels/modelcore/terminalsize.hpp
@@ -1,0 +1,12 @@
+/* -*- mode: c++; coding: utf-8; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4; show-trailing-whitespace: t -*- vim:fenc=utf-8:ft=cpp:et:sw=4:ts=4:sts=4
+ */
+
+#pragma once
+
+#include <utility>
+
+namespace Feel
+{
+std::pair<int,int> get_terminal_size();
+
+} // Feel
